requirements.txt
alembic==1.7.4
anyio==3.3.4
asgiref==3.4.1
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.8.6
pydantic==1.8.2
python-dotenv==0.19.1
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.25
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
